requirements.txt
agno
openai
pandas
duckdb
plotly
matplotlib
streamlit
python-dotenv
